sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docflow-client"
version = "0.2.0"
description = "Multi-format document conversion SDK optimized for RAG systems"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DocFlow Team", email = "docflow@example.com"}
]
keywords = ["markdown", "pdf", "docx", "excel", "csv", "rag", "converter", "docflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
requires-python = ">=3.9"
dependencies = [
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
]

[project.optional-dependencies]
# PDF generation engines
weasyprint = ["weasyprint>=60.0"]
pdfkit = ["pdfkit>=1.0.0"]
playwright = ["playwright>=1.40.0"]

# PDF extraction
pdf = ["pdfminer.six>=20221105"]

# Document formats
docx = ["python-docx>=1.1.0"]
excel = ["openpyxl>=3.1.0"]
formats = [
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
]

# LLM providers for RAG
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

# Cloud storage
s3 = ["boto3>=1.28.0"]
gcs = ["google-cloud-storage>=2.10.0"]
azure = ["azure-storage-blob>=12.0.0"]
storage = [
    "boto3>=1.28.0",
    "google-cloud-storage>=2.10.0",
    "azure-storage-blob>=12.0.0",
]

# All PDF features
all-pdf = [
    "weasyprint>=60.0",
    "pdfminer.six>=20221105",
]

# RAG mode (recommended)
rag = [
    "pdfminer.six>=20221105",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
]

# Everything
all = [
    "weasyprint>=60.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "boto3>=1.28.0",
    "azure-storage-blob>=12.0.0",
]

# Development
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/xgaslan/docflow"
Documentation = "https://github.com/xgaslan/docflow/sdks/python"
Repository = "https://github.com/xgaslan/docflow"

[tool.setuptools.packages.find]
where = ["."]
include = ["docflow*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
